Minor code changes and fixes. Design change in root action.
