+++ src/main/java/hudson/plugins/jobConfigHistory/JobConfigHistoryConsts.java
@@ -62,4 +62,7 @@ public final class JobConfigHistoryConsts {
 	public static final String ID_FORMATTER = "yyyy-MM-dd_HH-mm-ss";
 
 	public static final int DEFAULT_MAX_ENTRIES_PER_PAGE = Integer.MAX_VALUE;
+
+	/** Amount of neighbors displayed around the current page number in the page navigation bar. Example (epsilon=2, currentPage=6): [0 ... 4 5 <b>6</b> 7 8  ... 204] */
+	public static final int PAGING_EPSILON = 2;
 }
